requirements.txt
asgiref==3.4.1
braintree==4.12.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Django==3.2.7
django-credit-cards==0.4.1
django-crispy-forms==1.12.0
django-payments==0.14.0
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
stripe==2.60.0
suds-jurko==0.6
urllib3==1.26.7
xmltodict==0.12.0
